Increase code coverage and remove out folder before running tests
tsc ignores renaming files and leaves previously transpiled
files in out folder. That leads to coverage reporting to hang
after mocha exits. Fix adds clean script to run before tests.
