+++ test/openshift/component.test.ts
@@ -7,7 +7,7 @@ import * as sinon from 'sinon';
 import * as shelljs from 'shelljs';
 import { TestItem } from './testOSItem';
 import { OdoImpl } from '../../src/odo';
-import { ToolsConfig } from '../../src/toolsConfig';
+import { ToolsConfig } from '../../src/tools';
 import { Component } from '../../src/openshift/component';
 import { Progress } from '../../src/util/progress';
